Extract duplicated constants, validation hook, and reminder utils (#230, #285, #234)
- contactConstants.ts: shared SOCIAL_TYPES, NON_OTHER_SOCIAL_TYPES, SOCIAL_META, localToday()
- useContactFieldValidation: single hook for warning state + checkEmailBlur/checkPhoneBlur; replaces 5 duplicated state vars and 2 duplicated async functions in AddContactModal and GlobalTab
- logShared: export fmtReminderDate and sortReminders; remove byte-for-byte duplicates in GlobalRemindersSection and ProjectTab.RemindersSection

Closes #230
Closes #285
Closes #234
